fix: remove sharding circular dependency
